"Freeze" int types to fix some unsafe `@pure`s.
Require users to manually register new integral backing types for
FixedDecimal in order to see the best performance w/ that type.

It's super hacky, but to register the type, we re-define all the `@pure`
functions, so that they are defined _after_ the method tables have been
modified, so that they won't see any changes to those global mutable
state.

In exchange for this performance improvement, users commit to not
changing the definitions for the Base integer interface functions used
by FD, or suffer undefined behavior from the `@pure` functions.
If they want to change those methods, they must re-freeze the type after
changing the definitions.